Initial build of a more structured model for this system.
We now store and manage a list of bus stops locally (from the TfL api), we'll soon start streaming all bus data available and associating it with this, and then start doing location lookups on that all by ourselves: this should be quicker, and much more fun :D

Location querying is only semi-optimised for now: fast enough to appear near-instant with manual lookups, down from ~1 second naive, but probably not good enough to scale.

Note also: start of an effort to add unittests, here in doctest format.